+++ client/js/src/client.ts
@@ -756,7 +756,11 @@ export function api_factory(
 									}
 								}
 							};
-						} else if (protocol == "sse_v1" || protocol == "sse_v2") {
+						} else if (
+							protocol == "sse_v1" ||
+							protocol == "sse_v2" ||
+							protocol == "sse_v2.1"
+						) {
 							// latest API format. v2 introduces sending diffs for intermediate outputs in generative functions, which makes payloads lighter.
 							fire_event({
 								type: "status",
@@ -850,7 +854,10 @@ export function api_factory(
 													endpoint: _endpoint,
 													fn_index
 												});
-												if (data && protocol === "sse_v2") {
+												if (
+													data &&
+													(protocol === "sse_v2" || protocol === "sse_v2.1")
+												) {
 													apply_diff_stream(event_id!, data);
 												}
 											}