+++ client/python/gradio_client/utils.py
@@ -246,10 +246,12 @@ class Communicator:
 ########################
 
 
-def is_http_url_like(possible_url: str) -> bool:
+def is_http_url_like(possible_url) -> bool:
     """
-    Check if the given string looks like an HTTP(S) URL.
+    Check if the given value is a string that looks like an HTTP(S) URL.
     """
+    if not isinstance(possible_url, str):
+        return False
     return possible_url.startswith(("http://", "https://"))
 
 
@@ -390,7 +392,7 @@ async def get_pred_from_sse_v1_v2(
     cookies: dict[str, str] | None,
     pending_messages_per_event: dict[str, list[Message | None]],
     event_id: str,
-    protocol: Literal["sse_v1", "sse_v2"],
+    protocol: Literal["sse_v1", "sse_v2", "sse_v2.1"],
 ) -> dict[str, Any] | None:
     done, pending = await asyncio.wait(
         [
@@ -510,7 +512,7 @@ async def stream_sse_v1_v2(
     helper: Communicator,
     pending_messages_per_event: dict[str, list[Message | None]],
     event_id: str,
-    protocol: Literal["sse_v1", "sse_v2"],
+    protocol: Literal["sse_v1", "sse_v2", "sse_v2.1"],
 ) -> dict[str, Any]:
     try:
         pending_messages = pending_messages_per_event[event_id]
@@ -546,10 +548,10 @@ async def stream_sse_v1_v2(
                     log=log_message,
                 )
                 output = msg.get("output", {}).get("data", [])
-                if (
-                    msg["msg"] == ServerMessage.process_generating
-                    and protocol == "sse_v2"
-                ):
+                if msg["msg"] == ServerMessage.process_generating and protocol in [
+                    "sse_v2",
+                    "sse_v2.1",
+                ]:
                     if pending_responses_for_diffs is None:
                         pending_responses_for_diffs = list(output)
                     else:
@@ -621,6 +623,20 @@ def apply_diff(obj, diff):
 ########################
 # Data processing utils
 ########################
+
+
+def upload_file(
+    file_path: str,
+    upload_url: str,
+    headers: dict[str, str] | None = None,
+    cookies: dict[str, str] | None = None,
+):
+    with open(file_path, "rb") as f:
+        files = [("files", (Path(file_path).name, f))]
+        r = httpx.post(upload_url, headers=headers, cookies=cookies, files=files)
+    r.raise_for_status()
+    result = r.json()
+    return result[0]
 
 
 def download_file(
@@ -898,13 +914,18 @@ def get_type(schema: dict):
         raise APIInfoParseError(f"Cannot parse type for {schema}")
 
 
-OLD_FILE_DATA = "Dict(path: str, url: str | None, size: int | None, orig_name: str | None, mime_type: str | None)"
-FILE_DATA = "Dict(path: str, url: str | None, size: int | None, orig_name: str | None, mime_type: str | None, is_stream: bool)"
+FILE_DATA_FORMATS = [
+    "Dict(path: str, url: str | None, size: int | None, orig_name: str | None, mime_type: str | None)",
+    "Dict(path: str, url: str | None, size: int | None, orig_name: str | None, mime_type: str | None, is_stream: bool)",
+    "Dict(path: str, url: str | None, size: int | None, orig_name: str | None, mime_type: str | None, is_stream: bool, meta: Dict())",
+]
+
+CURRENT_FILE_DATA_FORMAT = FILE_DATA_FORMATS[-1]
 
 
 def json_schema_to_python_type(schema: Any) -> str:
     type_ = _json_schema_to_python_type(schema, schema.get("$defs"))
-    return type_.replace(FILE_DATA, "filepath")
+    return type_.replace(CURRENT_FILE_DATA_FORMAT, "filepath")
 
 
 def _json_schema_to_python_type(schema: Any, defs) -> str:
@@ -980,7 +1001,7 @@ def _json_schema_to_python_type(schema: Any, defs) -> str:
         raise APIInfoParseError(f"Cannot parse schema {schema}")
 
 
-def traverse(json_obj: Any, func: Callable, is_root: Callable) -> Any:
+def traverse(json_obj: Any, func: Callable, is_root: Callable[..., bool]) -> Any:
     if is_root(json_obj):
         return func(json_obj)
     elif isinstance(json_obj, dict):
@@ -999,25 +1020,55 @@ def traverse(json_obj: Any, func: Callable, is_root: Callable) -> Any:
 
 def value_is_file(api_info: dict) -> bool:
     info = _json_schema_to_python_type(api_info, api_info.get("$defs"))
-    return FILE_DATA in info or OLD_FILE_DATA in info
+    return any(file_data_format in info for file_data_format in FILE_DATA_FORMATS)
 
 
-def is_filepath(s):
-    return isinstance(s, str) and Path(s).exists()
+def is_filepath(s) -> bool:
+    """
+    Check if the given value is a valid str or Path filepath on the local filesystem, e.g. "path/to/file".
+    """
+    return isinstance(s, (str, Path)) and Path(s).exists() and Path(s).is_file()
 
 
-def is_url(s):
-    return isinstance(s, str) and is_http_url_like(s)
+def is_file_obj(d) -> bool:
+    """
+    Check if the given value is a valid FileData object dictionary in versions of Gradio<=4.20, e.g.
+    {
+        "path": "path/to/file",
+    }
+    """
+    return isinstance(d, dict) and "path" in d and isinstance(d["path"], str)
 
 
-def is_file_obj(d):
-    return isinstance(d, dict) and "path" in d
-
-
-def is_file_obj_with_url(d):
+def is_file_obj_with_meta(d) -> bool:
+    """
+    Check if the given value is a valid FileData object dictionary in newer versions of Gradio
+    where the file objects include a specific "meta" key, e.g.
+    {
+        "path": "path/to/file",
+        "meta": {"_type: "gradio.FileData"}
+    }
+    """
     return (
-        isinstance(d, dict) and "path" in d and "url" in d and isinstance(d["url"], str)
+        isinstance(d, dict)
+        and "path" in d
+        and isinstance(d["path"], str)
+        and "meta" in d
+        and d["meta"].get("_type", "") == "gradio.FileData"
     )
+
+
+def is_file_obj_with_url(d) -> bool:
+    """
+    Check if the given value is a valid FileData object dictionary in newer versions of Gradio
+    where the file objects include a specific "meta" key, and ALSO include a "url" key, e.g.
+    {
+        "path": "path/to/file",
+        "url": "/file=path/to/file",
+        "meta": {"_type: "gradio.FileData"}
+    }
+    """
+    return is_file_obj_with_meta(d) and "url" in d and isinstance(d["url"], str)
 
 
 SKIP_COMPONENTS = {
@@ -1034,3 +1085,16 @@ SKIP_COMPONENTS = {
     "interpretation",
     "dataset",
 }
+
+
+def file(filepath_or_url: str | Path):
+    s = str(filepath_or_url)
+    data = {"path": s, "meta": {"_type": "gradio.FileData"}}
+    if is_http_url_like(s):
+        return {**data, "orig_name": s.split("/")[-1], "url": s}
+    elif Path(s).exists():
+        return {**data, "orig_name": Path(s).name}
+    else:
+        raise ValueError(
+            f"File {s} does not exist on local filesystem and is not a valid URL."
+        )