+++ js/app/src/lite/custom-element/index.ts
@@ -1,6 +1,10 @@
 // NOTE: We should only import the types from ".." to avoid the circular dependency of implementations,
 // which causes repeated executions of the ".." module in †he dev mode and can lead to multiple instances of the dev app.
-import type { create as createLiteAppFunc, Options } from "..";
+import type {
+	create as createLiteAppFunc,
+	Options,
+	GradioAppController
+} from "..";
 
 interface GradioComponentOptions {
 	info: Options["info"];
@@ -40,22 +44,31 @@ export function bootstrap_custom_element(
 	}
 
 	class GradioLiteAppElement extends HTMLElement {
-		constructor() {
-			super();
+		controller: GradioAppController | null = null;
 
-			const gradioComponentOptions = this.parseGradioComponentOptions();
-			const gradioLiteAppOptions = this.parseGradioLiteAppOptions();
+		connectedCallback(): void {
+			// At the time of connectedCallback, the child elements of the custom element are not yet parsed,
+			// so we need to defer the initialization to the next frame.
+			// Ref: https://stackoverflow.com/q/70949141/13103190
+			window.requestAnimationFrame(() => {
+				const gradioComponentOptions = this.parseGradioComponentOptions();
+				const gradioLiteAppOptions = this.parseGradioLiteAppOptions();
 
-			this.innerHTML = "";
+				this.innerHTML = "";
 
-			create({
-				target: this, // Same as `js/app/src/main.ts`
-				code: gradioLiteAppOptions.code,
-				requirements: gradioLiteAppOptions.requirements,
-				files: gradioLiteAppOptions.files,
-				entrypoint: gradioLiteAppOptions.entrypoint,
-				...gradioComponentOptions
+				this.controller = create({
+					target: this, // Same as `js/app/src/main.ts`
+					code: gradioLiteAppOptions.code,
+					requirements: gradioLiteAppOptions.requirements,
+					files: gradioLiteAppOptions.files,
+					entrypoint: gradioLiteAppOptions.entrypoint,
+					...gradioComponentOptions
+				});
 			});
+		}
+
+		disconnectedCallback(): void {
+			this.controller?.unmount();
 		}
 
 		parseGradioComponentOptions(): GradioComponentOptions {