+++ js/audio/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @gradio/audio
 
+## 0.9.6
+
+### Dependency updates
+
+- @gradio/button@0.2.25
+- @gradio/upload@0.7.7
+- @gradio/client@0.13.0
+- @gradio/wasm@0.8.0
+
 ## 0.9.5
 
 ### Patch Changes