+++ js/button/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @gradio/button
 
+## 0.2.25
+
+### Dependency updates
+
+- @gradio/upload@0.7.7
+- @gradio/client@0.13.0
+
 ## 0.2.24
 
 ### Patch Changes